Improved voice selection in NoteOn
Previously one key might be played simultaneously by more than
one voice and a releasing voice could be selected in
CMiniSynthesizer::NoteOn() to play a new key. Now it is ensured,
that only one voice is playing a key at a time. If it is still
releasing and will be re-pressed, the same voice will be restarted.