Fix function of VCF and VCA LFOs
The LFOs did not work properly for VCF and VCA. Make the VCF and VCA
LFO frequencies more fine grained too. LFO depth can be up to 100% now.
There is a version property added to the patch files, which has the
value 2 now. Parameters from old patch files (without this property)
will be corrected on load to get a similar sound as before (as far as
possible).

Issue #8