kernel: Provide null device as logger target
Logger target is the screen by default. Normally this shouldn't
be a problem, because log messages should not occur after GUI
initialization, but in case some spurious message is generated
(e.g. from USB), the GUI screen would be destroyed. This can be
prevented by directing the logger to the new null device. To do
this one has to create a file cmdline.txt on the SD card with this
contents in the first line:

logdev=null